Changed box-sizing
in day 7, KP enlightened us about box-sizing:border-box, and I wish I knew about it before. This makes sure that while we apply padding and border, the size does not overflow but stays within.

Although it won't have any effect on the current state of this project, I am still keeping it applied. Also changed the text-alignment of founders desc to justify because it looks better.