Modified the navbar
this time added a logo. It would be easier to just add the logo to the first nav-list, but KP insisted on adding the logo outside of the nav bar so that people using sound to go through the website do not have to stumble upon the home button twice while being on the nav bar.

Also learned about flex-grow, which when 1 allows an item in the flexbox to grow (is set to 0 on default)